+++ loama/playwright.config.ts
@@ -1,0 +1,57 @@
+import { defineConfig, devices } from '@playwright/test';
+
+/**
+ * See https://playwright.dev/docs/test-configuration.
+ */
+export default defineConfig({
+    testDir: './tests',
+    /* Run tests in files in parallel */
+    fullyParallel: false,
+    /* Fail the build on CI if you accidentally left test.only in the source code. */
+    forbidOnly: !!process.env.CI,
+    /* Retry on CI only */
+    retries: process.env.CI ? 2 : 0,
+    // Always use 1 worker so we don't create unwanted env's by running 2 tests at the same time
+    // It forces to only run 1 test in 1 browser at a time
+    workers: 1,
+    /* Reporter to use. See https://playwright.dev/docs/test-reporters */
+    reporter: 'html',
+    /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
+    use: {
+        trace: 'on-first-retry',
+        baseURL: 'http://localhost:4173'
+    },
+
+    /* Configure projects for major browsers */
+    projects: [
+        {
+            name: 'chromium',
+            use: { ...devices['Desktop Chrome'] },
+        },
+        //
+        // {
+        //     name: 'firefox',
+        //     use: { ...devices['Desktop Firefox'] },
+        // },
+        //
+        // {
+        //     name: 'webkit',
+        //     use: { ...devices['Desktop Safari'] },
+        // },
+    ],
+
+    webServer: [
+        {
+            command: 'yarn build && yarn preview',
+            url: 'http://localhost:4173',
+            timeout: 120 * 1000,
+            reuseExistingServer: !process.env.CI,
+        },
+        {
+            command: 'yarn pods:reset && yarn pods:start',
+            url: 'http://localhost:8080',
+            timeout: 120 * 1000,
+            reuseExistingServer: !process.env.CI,
+        }
+    ]
+});